+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 magic the better. Most of the people probably were not aware that `byebug`
 was overriding stuff like `pp` or `eval`. Only keeping `ps` as the single
 "enhanced evaluation" command.
+* `verbose` setting.
 
 ## 5.0.0
 ### Fixed